Sensor Factory and OpenNI2 Implementation done
More thoughts about device handling and enumeration.
It seems, that there is no way to check if two sensors are on the same US-Bus. A second enumeration on one bus will segfault.
Identification though seems to be best by URI match as long as devices aren't swapped around all the ports all the time.

Stored intrinsics seem to be a dead end, would be nice as initial values for calibration. Maybe rgb has something, or the values can be polled somehow...